chore: removed toggle switch + refactoring
