+++ site/src/app/components/pass-sport-breadcrumb/PassSportBreadCrumb.tsx
@@ -1,12 +1,11 @@
 'use client';
 
-import { usePathname } from 'next/navigation';
 import PassSportBreadcrumbStandard from '@/app/components/pass-sport-breadcrumb/PassSportBreadcrumbStandard';
 import PassSportBreadcrumbPro from '@/app/components/pass-sport-breadcrumb/PassSportBreadcrumbPro';
+import { useIsProVersion } from '@/app/hooks/use-is-pro-version';
 
 export default function PassSportBreadcrumb() {
-  const pathname = usePathname();
-  const isProVersion = pathname?.startsWith('/v2/pro/');
+  const isProVersion = useIsProVersion();
 
   return isProVersion ? <PassSportBreadcrumbPro /> : <PassSportBreadcrumbStandard />;
 }
