+++ site/src/app/hooks/use-is-pro-version.tsx
@@ -1,0 +1,7 @@
+import { usePathname } from 'next/navigation';
+
+export function useIsProVersion() {
+  const pathname = usePathname();
+
+  return pathname?.startsWith('/v2/pro/');
+}
